Move Action out of Common, don't use savedV0.
Compiler error was annoying me anyway.
