+++ Common/Common.vcxproj.filters
@@ -43,7 +43,6 @@
     <ClInclude Include="x64Emitter.h" />
     <ClInclude Include="ArmEmitter.h" />
     <ClInclude Include="ArmABI.h" />
-    <ClInclude Include="Action.h" />
     <ClInclude Include="FixedSizeUnorderedSet.h" />
   </ItemGroup>
   <ItemGroup>
@@ -72,7 +71,6 @@
     <ClCompile Include="x64Emitter.cpp" />
     <ClCompile Include="ArmEmitter.cpp" />
     <ClCompile Include="ArmABI.cpp" />
-    <ClCompile Include="Action.cpp" />
     <ClCompile Include="ThunkArm.cpp" />
   </ItemGroup>
   <ItemGroup>
